[BugFix] Fixing everything to get the project working off of a fresh pull
Mega submit to fix errors preventing the engine from starting, including :
- Adding missing #includes
- Resolving missing symbols by renaming, moving, or deleting symbols
- Correcting paths, variable names
- Fixed typos, removed unnecessary characters, fixed several warnings, alphabetized a few things, changed ordering of STL and library #includes
- Managed library files by moving them or deleting duplicates in wrong directories
- Added a script to copy libraries to the root directory, as well as a delete libraries line in the clean up bat file

+++ Source/Core/Events/EventManager.cpp
@@ -44,7 +44,7 @@ namespace QwerkE {
 
     void EventManager::ProcessEvents()
     {
-        int size = m_EventList.size();
+        int size = (int)m_EventList.size();
 
         for (int i = 0; i < size; i++)
         {
